+++ compiler_gym/leaderboard/__init__.py
@@ -1,0 +1,17 @@
+# Copyright (c) Facebook, Inc. and its affiliates.
+#
+# This source code is licensed under the MIT license found in the
+# LICENSE file in the root directory of this source tree.
+"""This package contains modules that can be used for preparing leaderboard
+submissions.
+
+We provide `leaderboards
+<https://github.com/facebookresearch/CompilerGym#leaderboards>`_ to track the
+performance of user-submitted algorithms on compiler optimization tasks. The
+goal of the leaderboards is to provide a venue for researchers to promote their
+work, and to provide a common framework for evaluating and comparing different
+approaches. We accept submissions to the leaderboards through pull requests, see
+`here
+<https://facebookresearch.github.io/CompilerGym/contributing.html#leaderboard-submissions>`_
+for instructions.
+"""