+++ tests/version_test.py
@@ -8,7 +8,9 @@ import pkg_resources
 import pytest
 
 import compiler_gym
+from compiler_gym.util.runfiles_path import runfiles_path
 from packaging import version
+from tests.pytest_plugins.common import bazel_only
 from tests.test_main import main
 
 # Marker to skip a test if running under bazel.
@@ -30,6 +32,14 @@ def test_version_dunder_format():
 @install_test
 def test_setuptools_version():
     version = pkg_resources.require("compiler_gym")[0].version
+    assert version == compiler_gym.__version__
+
+
+@bazel_only
+def test_expected_version():
+    """Test that embedded compiler gym version matches VERSION file."""
+    with open(runfiles_path("VERSION")) as f:
+        version = f.read().strip()
     assert version == compiler_gym.__version__